+++ auth/tokenstorage.py
@@ -1,9 +1,8 @@
 from datetime import datetime, timedelta, timezone
 
-from services.redis import redis
-from auth.validations import AuthInput
-
 from auth.jwtcodec import JWTCodec
+from auth.validations import AuthInput
+from services.redis import redis
 from settings import ONETIME_TOKEN_LIFE_SPAN, SESSION_TOKEN_LIFE_SPAN