+++ services/db.py
@@ -6,7 +6,17 @@ import warnings
 from typing import Any, Callable, Dict, TypeVar
 
 import sqlalchemy
-from sqlalchemy import JSON, Column, Engine, Integer, create_engine, event, exc, func, inspect
+from sqlalchemy import (
+    JSON,
+    Column,
+    Engine,
+    Integer,
+    create_engine,
+    event,
+    exc,
+    func,
+    inspect,
+)
 from sqlalchemy.orm import Session, configure_mappers, declarative_base
 from sqlalchemy.sql.schema import Table